Converted to MD and fixed error
Converted the README back to Markdown so to have an easier time with linking the badge with the RTD manual. Was having a hard time with restructured text.

Also fixed a small error in the text, `cross platform` --> `cross-platform` and updated python3 to python3.7 since I am using the new formatted string format new to python3.7. 

ie: 

> world = 'World'
> print(f'Hello {world}')