Renamed configr its class Configr + More
I renamed the module configr.py to config.py and its class Configr to
Config to reduce the amount of typos when typing the module or class
names.

This branch was originally to deal with a module problem with Sphinx make
but it morphed to a lot more. The problem is still on going and has not
been resolved. See GH-25.

Other Changes
------------------

- Changed theme to default
- Made sure all modules begin with Added if __name__ == '__main__':
- Updated test to new config module
    - Module configr was changed to config. Updated the test file and
    tests to reflect this change.
    - Converted mock obj to use config instead of configr
- Fixed issue with import
    - Module test/test_tchelper had an issue with the way it imported
    Config
- Renamed a test to prevent error msg
    - test/test_tchelper.py was throwing up an error due to not finding
    module config. I renamed it to do_not_test_tchelper.py to avoid the
    error msg while I work on the issue.