Removed GUI *.py files from gui/ and more
The UI files no longer have to be converted to .py files in PySide2.
Check [Loading UI Files in Qt for Python](https://www.blog.pythonlibrary.org/2018/05/30/loading-ui-files-in-qt-for-python/)
for more information. Close #22

The config.ini file was replaced with config.yaml. From this point on I
will be using PyYaml instead of ConfigParser. ConfigParser proved to
be cumbersome while yaml files also are a lot more extendable and
cleaner to read.